+++ AtlasKit/Sources/Chart/ChartDataService.swift
@@ -1,0 +1,14 @@
+//
+//  ChartDataService.swift
+//  AtlasKit
+//
+//  Created by Jannic Marcon on 25.08.2025.
+//
+import AtlasCore
+
+struct ChartDataService {
+    public func getAllKlineCharts() -> [Chart] {
+        //TODO: make function
+        return []
+    }
+}
